ant move

+++ src/leaf.java
@@ -23,7 +23,7 @@ public class leaf extends gameObject implements food{
     }
 
     public void Render(Graphics var1) {
-        var1.drawImage(image,x,y,null);
+        var1.drawImage(image,(int)x,(int)y,null);
        // var1.fillRect(x, y, 20, 20);
     }
     public Rectangle getBounds() {
